Moved world position logic to a separate file.
